Make marketing page the actual home page
- Replace src/app/page.tsx with full marketing content
- Remove app/(marketing) route group directory
- Keep all features: hero, CTAs with URL preservation, how-it-works section
- Maintain generateMetadata for SEO optimization
